requirements.txt
asgiref==3.2.10
bson==0.5.8
dataclasses==0.6
Django==2.2.13
django-filter==2.3.0
djangorestframework==3.11.0
djongo==1.3.2
importlib-metadata==1.6.1
Markdown==3.2.2
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.2.4
zipp==3.1.0
